+++ clinica/converters/adni_to_bids/modality_converters/_fmap.py
@@ -24,7 +24,7 @@ def convert_fmap(
     subjects: Iterable[str],
     force_new_extraction: bool = False,
     n_procs: int = 1,
-    pet_processing_step: int = 2,
+    **kwargs,
 ):
     """Convert field map images of ADNI into BIDS format.
