+++ clinica/converters/adni_to_bids/modality_converters/_fmri.py
@@ -17,7 +17,7 @@ def convert_fmri(
     force_new_extraction: bool = False,
     n_procs: int = 1,
     convert_multiband: bool = True,
-    pet_processing_step: int = 2,
+    **kwargs,
 ):
     """Convert fMR images of ADNI into BIDS format.
