+++ clinica/converters/adni_to_bids/modality_converters/_pet_utils.py
@@ -2,12 +2,14 @@ from enum import Enum
 
 import pandas as pd
 
-from clinica.converters.adni_to_bids._utils import ADNIModalityConverter
+from clinica.converters.adni_to_bids._utils import (
+    ADNIModalityConverter,
+    ADNIPETPreprocessingStep,
+)
 from clinica.utils.stream import cprint
 
 __all__ = [
     "get_images_pet",
-    "ADNIPETPreprocessingStep",
     "ADNITracer",
     "define_pet_processing_step_with_tracer",
 ]
@@ -19,35 +21,6 @@ class ADNITracer(str, Enum):
     PIB = "PIB"
     FDG = "FDG"
     AV1451 = "AV1451"
-
-
-class ADNIPETPreprocessingStep(Enum):
-    """ADNI preprocessing steps."""
-
-    STEP0 = "ADNI Brain PET: Raw"
-    STEP1 = "Co-registered Dynamic"
-    STEP2 = "Co-registered, Averaged"
-    STEP3 = "Coreg, Avg, Standardized Image and Voxel Size"
-    STEP4_8MM = "Coreg, Avg, Std Img and Vox Siz, Uniform Resolution"
-    STEP4_6MM = "Coreg, Avg, Std Img and Vox Siz, Uniform 6mm Res"
-
-    @classmethod
-    def from_step_value(cls, step_value: int):
-        """Accept step specification in raw integer (0, 1, ..., 5)."""
-        error_msg = (
-            f"Step value {step_value} is not a valid ADNI preprocessing step value."
-            f"Valid values are : \n"
-            f"{"\n".join([f"{list(ADNIPETPreprocessingStep).index(step)} : {step.value}" for step in list(ADNIPETPreprocessingStep)])}."
-        )
-        if step_value != int(step_value):
-            raise ValueError(error_msg)
-        if 0 <= step_value <= 5:
-            if step_value == 4:
-                return cls.STEP4_8MM
-            if step_value == 5:
-                return cls.STEP4_6MM
-            return cls[f"STEP{step_value}"]
-        raise ValueError(error_msg)
 
 
 def define_pet_processing_step_with_tracer(
